Syntax/pyproject.toml
[project]
name = "syntax"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asgiref==3.8.1",
    "dj-rest-auth>=7.0.1",
    "django==5.1.5",
    "django-allauth>=65.3.1",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt>=5.4.0",
    "drf-spectacular>=0.28.0",
    "environ==1.0",
    "psycopg2-binary==2.9.10",
    "python-dotenv==1.0.1",
    "requests>=2.32.3",
    "sqlparse==0.5.3",
    "typing-extensions==4.12.2",
    "tzdata==2025.1",
]
